Adds Scrollbar component
No event listeners yet, so clicking and dragging the scrollbar is not possible yet

Horizontal and vertical scrollbar, automatic sizing + visibility depending on the size of the provided "content window".

You initialize the scrollbar like so:
`local Jello = loadfile("Jello.lua")();
local Scrollbar = Jello.Components.Scrollbar(ParentWindow);`
where ParentWindow is a terminal that the scrollbar will make a window inside of. This window will be treated as if it was a dedicated window for the scrollbar (don't use it for something else).
To set the content the scrollbar .. well .. scrolls .. call Scrollbar.SetContent(window) where window is the window you wish to scroll around. If empty, the scrollbar will create a window the exact size of the parent window (no scrollbars visible). If the window passed is bigger in one and/or the other direction, the appropriate scrollbars will become visible and move the window around.

The "content" window lives inside the "contentContainer" window (which is the exact size of the parent window - the width/height of the visible scrollbars (i.e. making space for them). Doing this reduces/removes any jitter from moving the content window around since it'll no longer draw over the scrollbars.